requirements.txt
accelerate==0.25.0
addict==2.4.0
altair==5.2.0
annotated-types==0.6.0
ansi2html==1.9.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astor==0.8.1
axial_positional_embedding==0.2.1
bids==0.0
bids-validator==1.13.1
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
braindecode==0.8.1
cachetools==5.3.2
catalogue==2.0.10
cfgv==3.4.0
clip==0.2.0
cloudpathlib==0.16.0
CoLT5-attention==0.11.1
confection==0.1.4
ConfigArgParse==1.7
contexttimer==0.3.3
contourpy==1.1.1
cycler==0.12.0
cymem==2.0.8
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
decord==0.6.0
diffusers==0.24.0
distlib==0.3.8
docker-pycreds==0.4.0
docopt==0.6.2
docstring-inheritance==2.1.0
easydict==1.9
einops==0.8.0
evdev==1.6.1
fairscale==0.4.4
filelock==3.13.1
Flask==3.0.3
fonttools==4.43.0
formulaic==0.5.2
fsspec==2023.12.2
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h5py==3.10.0
hf_transfer==0.1.8
huggingface-hub==0.19.4
identify==2.5.33
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.11.0
iniconfig==2.0.0
interface-meta==1.3.0
iopath==0.1.10
itsdangerous==2.1.2
joblib==1.3.2
kaggle==1.5.16
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
llvmlite==0.41.0
local-attention==1.9.15
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.8.0
mdurl==0.1.2
mne==1.5.1
mne-bids==0.13
moviepy==1.0.3
munkres==1.1.4
murmurhash==1.0.10
nbformat==5.7.0
networkx==3.2.1
Nib==0.5.7
nibabel==5.3.2
num2words==0.5.13
numpy==1.26.4
open_clip_torch==2.29.0
plotly==5.24.1
pluggy==1.5.0
pooch==1.8.2
portalocker==3.0.0
product_key_memory==0.2.11
proglog==0.1.10
protobuf==5.29.1
pybids==0.17.2
pydantic==2.10.3
pydantic_core==2.27.1
PySide6==6.8.0.2
pytest==8.3.4
python-slugify==8.0.4
pytorch-msssim==1.0.0
reformer-pytorch==1.4.4
regex==2024.11.6
retrying==1.3.4
safetensors==0.4.5
sentry-sdk==2.19.0
setproctitle==1.3.4
sh==2.1.0
shiboken6==6.8.0.2
skorch==1.0.0
smmap==5.0.1
srsly==2.4.8
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
text-unidecode==1.3
timm==1.0.12
tokenizers==0.13.3
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchinfo==1.8.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.28.0
triton==3.1.0
universal_pathlib==0.2.5
wandb==0.19.0
Werkzeug==3.0.6
wrapt==1.17.0
zstandard==0.23.0
